Allow ladder to support setting replicas to 0
This removes the restriction that the lowest number of replicas that a
ladder can configure is 1. The linear mode already supports scaling
to zero with the config option preventSinglePointFailure. I believe
having an explicit safety option like this is not as necessary for
ladder as the zero case has to be explicitly enumerated.

This also removes the validation that the cores or nodes in the ladder
are 1 or greater, as 0 is a valid value for _schedulable_ nodes or cores
(especially when filtering by node labels). This is likely not strictly
necessary, but it makes the boundary condition somewhat more clear.